抽离配置文件，以进行持久化配置，防止git pull后导致的自定义配置丢失
