+++ config.default.yaml
@@ -1,0 +1,26 @@
+env:
+  conda_base_path: "C:/Users/YourUsername/Miniconda3"
+  conda_env_name: "hifisampler"
+  python_script_path: "./hifiserver.py"
+
+model:
+  vocoder_path: "./pc_nsf_hifigan_44.1k_hop512_128bin_2025.02/model.ckpt"
+  model_type: "ckpt"
+  hnsep_model_path: "./hnsep/vr/model.pt"
+
+audio:
+  sample_rate: 44100
+  win_size: 2048
+  hop_size: 512
+  origin_hop_size: 128
+  n_fft: 2048
+  n_mels: 128
+  mel_fmin: 40
+  mel_fmax: 16000
+
+processing:
+  wave_norm: false
+  loop_mode: false
+  peak_limit: 1.0
+  fill: 6
+
